weather project/requirements.txt
django==3.1.4
django-bootstrap4==2.3.1
requests==2.25.1
django-braces==1.14.0
Pillow==8.1.0